Nakki
This is based on Douglas McKechie's Wheel of Fortune available on GitHub. The purpose of it would be randomize people to win a prize or get a nakki (a task). Nakki.html is basically Douglas's code where you can see how the wheel should work (the checkboxes won't do anything). Nakki3.html is an attempt to make the toggle Simo and toggle Jonne buttons in the top of the screen working, idea being that the array theWheel.segments should update according to the clicks. This is required so that the list of people in the wheel of fortune can be edited if somebody is not present at the moment. Nakki3.html is bit of a mess, I have tried the "banana, apple, peach" array to see how the array should work and the theWheel.segments array looks different from that. Using dataArray to push the data in first and then pushing the same data to theWheel.segments.